kva: try to allocate a shared mem first
If failing to try to get a shared mem, allocating may be fail, too.
Because if some process succeeded to allocate, then others may fail.

However, if failing to try to allocate a shared mem, getting may not be
fail. Because this means some process allocated already if memory is
enough.

    modified:   kva.c